Switch from bits to byte for the argument
It doesn't make much sense to take the argument as a number of bits,
if the utility itself cannot represent values which are not a multiple
of 8. Instead, just accept a number of bytes and simplify the code.
It also makes it much clearer that we're dealing with a glorified static array.